requirements.txt
streamlit==1.35.0
Pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
google-generativeai==0.7.2
protobuf==4.25.3
opencv-python-headless==4.9.0.80 